Add null checks for File.listFiles() to prevent possible NPEs
- File.list() and File.listFiles() can return null if not a directory

+++ com.archimatetool.editor/src/com/archimatetool/editor/WorkbenchCleaner.java
@@ -217,12 +217,15 @@ public class WorkbenchCleaner {
     
     private static void deleteFolder(File folder) {
         if(folder.exists() && folder.isDirectory()) {
-            for(File file : folder.listFiles()) {
-                if(file.isFile()) {
-                    file.delete();
-                }
-                else if(file.isDirectory()) {
-                    deleteFolder(file);
+            File[] files = folder.listFiles();
+            if(files != null) {
+                for(File file : files) {
+                    if(file.isFile()) {
+                        file.delete();
+                    }
+                    else if(file.isDirectory()) {
+                        deleteFolder(file);
+                    }
                 }
             }
